[refactor] tui.py をパッケージ化してタブ実装を別ファイルに分割
単一の tui.py にまとまっていた実装をファイル単位で分割:

- tui/state.py      : TuiState / TuiResult / Tab ごとの sub-state / FIXED_ROWS
- tui/render.py     : pad_display / render_meta_table (共通描画ヘルパ)
- tui/tab.py        : TabView dataclass + noop / no_action デフォルト
- tui/issue_tab.py  : ISSUE_TAB の render / handler / load_journals 等
- tui/wiki_tab.py   : WIKI_TAB の render / handler / wiki ロード関数
- tui/app.py        : TABS レジストリ + run_issue_tui 本体 +
                      スクリーンダンプヘルパ
- tui/__init__.py   : 既存 import 互換のため TuiState / run_issue_tui
                      等を再エクスポート

タブ追加のフットプリントが個別ファイルに閉じ、tui/app.py は
TABS レジストリへの登録と run_issue_tui のみを持つ形になった。
振る舞いは同じ (task check 通過)。
