+++ src/redi/tui/__init__.py
@@ -1,0 +1,22 @@
+from redi.tui.app import dump_rendered_screen, run_issue_tui
+from redi.tui.state import (
+    IssueTabState,
+    TuiAction,
+    TuiPosition,
+    TuiResult,
+    TuiState,
+    TuiTab,
+    WikiTabState,
+)
+
+__all__ = [
+    "IssueTabState",
+    "TuiAction",
+    "TuiPosition",
+    "TuiResult",
+    "TuiState",
+    "TuiTab",
+    "WikiTabState",
+    "dump_rendered_screen",
+    "run_issue_tui",
+]
