+++ src/redi/tui/state.py
@@ -1,0 +1,48 @@
+from dataclasses import dataclass, field
+from typing import Literal
+
+TuiAction = Literal["update", "create", "comment"]
+TuiTab = Literal["issues", "wiki"]
+
+# 一覧/プレビューの外側にある固定行の合計 (タブバー + 罫線 + ステータスバー)。
+# Layout の HSplit に固定行を増減したらここも更新すること。
+FIXED_ROWS = 3
+
+
+@dataclass
+class TuiPosition:
+    offset: int = 0
+    cursor: int = 0
+
+
+@dataclass
+class TuiResult:
+    action: TuiAction
+    issue_id: str | None
+    position: TuiPosition
+
+
+@dataclass
+class IssueTabState:
+    offset: int = 0
+    cursor: int = 0
+    issues: list[dict] = field(default_factory=list)
+
+
+@dataclass
+class WikiTabState:
+    loaded: bool = False
+    pages: list[dict] = field(default_factory=list)
+    labels: list[str] = field(default_factory=list)
+    cursor: int = 0
+    texts: dict[str, str] = field(default_factory=dict)
+    error: str | None = None
+
+
+@dataclass
+class TuiState:
+    last_result: TuiResult | None = None
+    page_size: int = 0
+    tab: TuiTab = "issues"
+    issue_tab: IssueTabState = field(default_factory=IssueTabState)
+    wiki_tab: WikiTabState = field(default_factory=WikiTabState)
